create dynamic rendering for muscle Group, and dynamic route to dnymic exercise
