+++ components/Main/MuscleGroups/MuscleGroupItem/MuscleGroupItem.js
@@ -5,14 +5,16 @@ import { StyledMuscleGroupItem } from "./StyledMuscleGroupItem";
 const MuscleGroupItem = ({ muscle }) => {
   return (
     <StyledMuscleGroupItem>
-      <Link href="/exercises">
-        <Image
-          src="/images/abs.svg"
-          width={158}
-          height={80}
-          alt="add-fav icon"
-        />
-        <h3>{muscle}</h3>
+      <Link href={`/exercises/${muscle}`}>
+        <li>
+          <Image
+            src={`/images/${muscle}.svg`}
+            width={158}
+            height={80}
+            alt="muscle"
+          />
+          <h3>{muscle}</h3>
+        </li>
       </Link>
     </StyledMuscleGroupItem>
   );
